@@ -39,8 +39,8 @@
         "@0x/typescript-typings": "^5.2.0",
         "@0x/utils": "^6.4.3",
         "@0x/web3-wrapper": "^7.5.3",
-        "@ethereumjs/common": "^2.2.0",
-        "@ethereumjs/tx": "^3.1.3",
+        "@ethereumjs/common": "^2.3.1",
+        "@ethereumjs/tx": "^3.2.1",
         "@ledgerhq/hw-app-eth": "^4.3.0",
         "@ledgerhq/hw-transport-u2f": "4.24.0",
         "@types/hdkey": "^0.7.0",
@@ -60,8 +60,8 @@
     "devDependencies": {
         "@0x/ts-doc-gen": "^0.0.28",
         "@0x/tslint-config": "^4.1.4",
-        "@types/bip39": "^2.4.0",
         "@types/bn.js": "^4.11.0",
+        "@types/chai": "^4.2.19",
         "@types/ethereumjs-tx": "^1.0.1",
         "@types/hdkey": "^0.7.0",
         "@types/lodash": "4.14.104",
@@ -78,7 +78,7 @@
         "sinon": "^4.0.0",
         "tslint": "5.11.0",
         "typedoc": "~0.16.11",
-        "typescript": "4.2.2",
+        "typescript": "4.3.4",
         "webpack": "^4.20.2"
     },
     "optionalDependencies": {
